Fix: 삭제할 알람이 없는 경우의 오류 반환

+++ src/main/java/com/openbook/openbook/basicuser/service/BasicUserService.java
@@ -70,6 +70,9 @@ public class BasicUserService {
     @Transactional
     public void deleteAllAlarm(final Long userId) {
         User user = userService.getUserOrException(userId);
+        if(!alarmService.userHasReceivedAlarms(userId)) {
+            throw new OpenBookException(ErrorCode.ALARM_NOT_FOUND);
+        }
         alarmService.deleteAllReceiverAlarm(user.getId());
     }
